Add compression and multi-image TIFF analysis
Also make test files a bit easier to dig through.

Fixes #35 and #79
