+++ .pre-commit-config.yaml
@@ -5,7 +5,7 @@ repos:
     rev: v5.0.0
     hooks:
       - id: check-added-large-files
-        args: ['--maxkb=1024']
+        args: ['--maxkb=1500']
       - id: check-ast
       - id: check-case-conflict
       - id: check-merge-conflict
